requirements.txt
streamlit>=1.39
google-genai>=0.3.0
openai>=1.57.0
anthropic>=0.40.0
pyyaml>=6.0.2
datasets>=3.0.0
